Move job constants to auxiliary constants
