+++ app/src/main/java/de/tum/in/tumcampusapp/services/FillCacheService.java
@@ -8,7 +8,7 @@ import android.support.v4.app.JobIntentService;
 import de.tum.in.tumcampusapp.auxiliary.Utils;
 import de.tum.in.tumcampusapp.managers.CacheManager;
 
-import static de.tum.in.tumcampusapp.services.JobIdConstants.FILL_CACHE_SERVICE_JOB_ID;
+import static de.tum.in.tumcampusapp.auxiliary.Const.FILL_CACHE_SERVICE_JOB_ID;
 
 /**
  * Service used to fill caches in background, for faster/offline access
