--- app/src/main/java/de/tum/in/tumcampusapp/services/JobIdConstants.java
@@ -1,16 +1,0 @@
-package de.tum.in.tumcampusapp.services;
-
-/**
- * Created by ansel on 11/11/2017.
- */
-
-final class JobIdConstants {
-
-    static final int BACKGROUND_SERVICE_JOB_ID = 1000;
-    static final int SEND_MESSAGE_SERVICE_JOB_ID = 1001;
-    static final int SILENCE_SERVICE_JOB_ID = 1002;
-    static final int SEND_WIFI_SERVICE_JOB_ID = 1003;
-    static final int DOWNLOAD_SERVICE_JOB_ID = 1004;
-    static final int FILL_CACHE_SERVICE_JOB_ID = 1005;
-
-}
